web_app/requirements.txt
Flask==2.3.2
gspread==5.10.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
gunicorn==22.0.0
